trying to manually fix stuff
aight so, there was a bug that needed fixing back in like mid-April, and I decided to try to fix it myself. I could've waited for Nick Melhado to fix it and then simply clicked the "sync fork" button, but I was impatient. I did successfully fix it, just not in the exact same way, so now I'm trying to manually input his fixes to make my fork identical. Because where things currently stand, it appears that I am unable to merge any of his updates to my fork as long as there's a discrepancy like this. Hopefully this will fix that, but idk if it actually will or not.