Fixing a BIG oopsie
When I removed the line of, "Scrappy is in charge of kicking our asses every year", I also removed a "</ul>" line, and that messed EVERYTHING up. Only the opening paragraph was visible. Whoops. But we should be good now I hope?

(I also indented a bunch of lines for the sake of aesthetics in the code, which I expect is fine, but I feel obligated to mention that as well)